Wire the library sections' playlist refresh, and stop evaluating them twice
Two defects from the same corner of the admin page.

#2679 — saving a playlist from the Crate Digger panel never refreshed the
list. The refresh was written; it just never ran. Both library.js and
library-ai.js guarded the call on `window.loadPlaylists`, a global nothing in
www/ has ever assigned, so the guard was permanently false. The sections now
take a `reloadPlaylists` parameter, the seam PR #2677 established with
`initMixTab({ startGame, refreshStatus })` and `initMediaPlayers({
refreshStatus })`; admin.js hands them its own `loadStatus`.

#2680 — admin/sections/library.js was evaluated twice per page. wizard.js
imported it over the network while admin.min.js inlined it, so the browser
held two module instances with two sets of module-level state (and the same
for library-ai.js, admin/state.js, game-constants.js, providers.generated.js).
Resolved by moving wizard.js into the admin bundle; playlist-hub.js comes with
it because wizard.js imports it. A file is a bundle input or its own entry
point, never both.

The #2677 check compared admin.html's `<script src>` names against the esbuild
metafile, which only sees a duplicate the page names out loud. It now walks
each non-bundled script's whole import graph, which is how #2680 slipped past
it.

Closes #2679
Closes #2680
